requirements.txt
aicscytoparam==0.1.9
aicsimageio==4.11.0
aicspylibczi>3.0.0
aicsshparam==0.1.7
aiobotocore==2.4.2
aiohttp==3.8.5
aioitertools==0.11.0
aiosignal==1.3.1
aiosqlite==0.19.0
alembic==1.11.1
aniso8601==9.0.1
anndata==0.9.2
annotated-types==0.5.0
ansi2html==1.8.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
apprise==1.4.5
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asciitree==0.3.3
asgi-lifespan==2.1.0
astropy==5.2.2
asttokens==2.2.1
async-timeout==4.0.2
asyncpg==0.28.0
attrs==23.1.0
autograd==1.6.2
autopage==0.5.1
backcall==0.2.0
backoff==2.2.1
backports.zoneinfo;python_version<"3.9"
beautifulsoup4==4.12.2
bfio==2.3.0
bleach==6.0.0
blessed==1.20.0
blinker==1.6.2
bokeh==3.1.1
boto3==1.24.4
botocore==1.27.59
cachetools==5.3.1
camera-alignment-core==1.0.5
cellpose-omni==0.9.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
cliff==4.3.0
cloudpickle==2.2.1
cmaes==0.10.0
cmake==3.27.0
cmd2==2.4.3
colorama==0.4.6
colorcet==3.0.1
colorlog==6.7.0
comm==0.1.4
contourpy==1.1.0
coolname==2.2.0
coverage==7.2.7
croniter==1.4.1
cryptography==41.0.3
cycler==0.11.0
cyto_dl @ git+ssh://git@github.com/AllenCellModeling/cyto-dl.git@main
dask==2023.8.0
dask-jobqueue
databricks-cli==0.17.7
dateparser==1.1.8
dateutils==0.6.12
decorator==5.1.1
deepdiff==6.3.1
defusedxml==0.7.1
distributed
docker==6.1.3
docutils==0.20.1
e3nn==0.5.1
edt==2.3.1
einops==0.6.1
elementpath==4.1.5
entrypoints==0.4
escnn==1.0.11
exceptiongroup==1.1.2
executing==1.2.0
fastapi==0.100.1
fasteners==0.18
fastjsonschema==2.18.0
fastremap==1.13.5
filelock==3.12.2
fire==0.5.0
flake8==6.1.0
Flask==2.3.2
Flask-RESTful==0.3.10
fonttools==4.41.1
frozenlist==1.4.0
fsspec==2023.3.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.22.0
gorilla==0.4.0
greenlet==2.0.2
griffe==0.32.3
gunicorn==20.1.0
h11==0.14.0
h2==4.1.0
h5py==3.9.0
hpack==4.0.0
httpcore==0.17.3
httpx==0.24.1
hydra-colorlog==1.2.0
hydra-core==1.3.2
hydra-optuna-sweeper==1.2.0
hyperframe==6.0.1
idna==3.4
igraph==0.10.6
imagecodecs==2023.3.16
imageio==2.31.1
importlib-metadata==6.8.0
importlib-resources==6.0.0
iniconfig==2.0.0
inquirer==3.1.3
ipydatawidgets==4.3.2
ipykernel==6.25.0
ipympl==0.9.3
ipython==8.13.0
ipython-genutils==0.2.0
ipywidgets==7.8.0
itk-core==5.3.0
itk-filtering==5.3.0
itk-meshtopolydata==0.10.0
itk-numerics==5.3.0
itkwidgets==0.32.6
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.1
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.5
kiwisolver==1.4.4
kubernetes==27.2.0
lazy_loader==0.3
lie-learn==0.0.1.post1
lightning==2.0.6
lightning-cloud==0.5.37
lightning-utilities==0.9.0
lit==16.0.6
llvmlite==0.40.1
locket==1.0.0
lxml==4.9.3
mahotas==1.4.13
Mako==1.2.4
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mgen==1.2.1
mistune==3.0.1
mlflow==2.5.0
monai-weekly==1.3.dev2331
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
natsort==8.4.0
nbclient==0.8.0
nbconvert==7.7.3
nbformat==5.9.2
ncolor==1.2.1
nest-asyncio==1.5.7
networkx==3.1
notebook==6.4.12
numba==0.57.1
numcodecs==0.11.0
numpy==1.24.4
nvgpu==0.10.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
ome-types==0.4.1
ome-zarr==0.8.0
omegaconf==2.3.0
omnipose @ git+https://github.com/kevinjohncutler/omnipose.git@ce2e9aaca2716718b6a3f30c27a99d147135701c
opencv-python-headless==4.8.0.74
opt-einsum==3.3.0
opt-einsum-fx==0.1.4
optuna==2.10.1
ordered-set==4.1.0
orjson==3.9.2
ostat==0.2
packaging==23.1
pandas==1.5.3
pandocfilters==1.5.0
param==1.13.0
parso==0.8.3
partd==1.4.0
pathspec==0.11.2
pbr==5.11.1
peakdetect==1.2
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
pluggy==1.2.0
pooch==1.7.0
prefect==2.11.2
prefect-dask
prettytable==3.8.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==3.20.1
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py3nj==0.1.2
pyarrow==10.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.0
pycparser==2.21
pyct==0.5.0
pydantic==1.10.12
pydantic_core==2.3.0
pyerfa==2.0.0.3
pyflakes==3.1.0
Pygments==2.15.1
PyJWT==2.8.0
pymanopt==2.1.1
pyparsing==3.0.9
pyperclip==1.8.2
pyshtools==4.10.3
pytest==7.4.0
pytest-cov==4.1.0
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
python-slugify==8.0.1
pytorch-ranger==0.1.1
pytz==2023.3
pytzdata==2020.1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
querystring-parser==1.2.4
readchar==4.0.5
referencing==0.30.0
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
resource-backed-dask-array==0.1.0
rich==13.5.2
rpds-py==0.9.2
rsa==4.9
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
s3fs==2023.3.0
s3transfer==0.6.1
scikit-image==0.19.2
scikit-learn==1.3.0
scipy==1.10.1
Send2Trash==1.8.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==1.4.49
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.27.0
starsessions==1.3.0
stevedore==5.1.0
sympy==1.12
tabulate==0.9.0
tblib==2.0.0
termcolor==2.3.0
terminado==0.17.1
text-unidecode==1.3
texttable==1.6.7
threadpoolctl==3.2.0
tifffile==2023.7.10
timelapsetracking @ git+ssh://git@github.com/aics-int/aics-timelapse-tracking.git@496c1441d373b281eefba6cc6f188e1f4cbd020c#egg=timelapsetracking
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torch-optimizer==0.3.0
torchmetrics==1.0.1
torchserve==0.8.1
torchvf==0.1.3
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.6.0
traittypes==0.2.1
triton==2.0.0
typer==0.9.0
typing_extensions==4.7.1
tzlocal==5.0.1
universal-pathlib==0.0.24
urllib3==1.26.16
uvicorn==0.23.2
vtk==9.2.6
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.6.1
websockets==11.0.3
Werkzeug==2.3.6
widgetsnbextension==3.6.5
wrapt==1.15.0
xarray==2023.1.0
xmlschema==2.4.0
xsdata==23.7
xyzservices==2023.7.0
yarl==1.9.2
zarr==2.16.0
zict==3.0.0
zipp==3.16.2
zstandard==0.21.0
